Encoder / Decoder framework
Started laying out the ground works of how the program is going to
encode and decode data. There are two functions, these return byte
buffers. This way the main function is just able to call these functions
which will handle the different modes of encoding and decoding.
I also added "enc_dec_mode(s)" variables which allows the user to
specify how data should be encoded and decoded (i.e. the audio is from a
commodore datasette and is not program standard).