It works!
I can encode and decode data. The reason the buffer was behaving
strangely was because I forgot to multiply the allocation
amount by the size of a float. Thus: buffer overflow.
I also decreased the speed at which time passes, making it finer, so now
frequency modulated encodings have a much smoother tone and waveform.