Seamless FM encoding
Added some code that would align the end of each bit to
zero. This way the encoder generates a nice smooth line.
Although not for pulse encoding. Next things to do are to make this code
work for pulse encoding and to make it work with a variable cycle count.