Backend reliance
I have obliterated the wrapper functions in favor of having the front
end interface with the backend directly. This way the backend has all of
the control. There are some issues with my backend that I need to fix.
Such as when the decoding data, there is apparently no data put into the
buffer (that is written to the file at least).
Overall, this is a much cleaner setup, and much more free setup than
having a front, middle, and multiple backends.