Fixed decoding
A pointer to a buffer had to be passed to the function not just the
buffer itself. It is a little bit strange, but it works. The code for
decoding pulse modulated files had nothing wrong with it, but there were
some logical errors with the code to do it for frequency modulated
files.
I am currently realizing many simplifications that can be made to my
code. This commit is just to have a reference point of everything
working.